// src/app/models/common/dataview-model.ts
import { DataviewListModel } from "./dataview-list-model";

export class DataviewModel {
    
    dataviewlist: DataviewListModel[];
    linkTitleIn:boolean;
    codModal:number;
    imagePathEmpaque: string ="assets/layout/images/dashboard/paquete.png";
    codModalImg:number;
    imagePathAvatar: string ="assets/layout/images/avatar-unisex.jpg";
    indPerson?: number
    
   
    //typePerson:number
}